fix(frontend): split PreorderCountdown helpers for react-refresh lint
Move pure countdown helpers into PreorderCountdown.utils.ts so the
component module only exports the component, satisfying ESLint
react-refresh rules and fixing CI.
